Added a utils package
Added
- date_utils.py to include datetime specific helper functinos
- Updated unit tests to reflect the updated package
